pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "django-fixedassets-ve"
version = "0.1.0"
description = "Fixed assets register with SLM/WDV depreciation, revaluation, disposal and accounting integration for Django."
readme = "README.md"
requires-python = ">=3.11"
authors = [{name = "Verified Enquiries / Tushar K.", email = "support@verifiedenquiries.com"}]
license = {text = "MIT"}
dependencies = [
  "Django>=5.0",
  "djangorestframework>=3.15",
  "django-finacc>=0.1.0",
]
classifiers = [
  "Framework :: Django",
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
]

[tool.setuptools]
include-package-data = true
packages = [
  "fixedassets", "fixedassets.utils", "fixedassets.models", "fixedassets.posting", "fixedassets.api", "fixedassets.mgmt.commands"
]